+++ SPTracer/src/Model/Color/ScalarColor.h
@@ -9,11 +9,11 @@ namespace SPTracer
 	class ScalarColor : public Color
 	{
 	public:
-		ScalarColor(double amplitude);
-		double GetAmplitude(double waveLength) const override;
+		ScalarColor(float amplitude);
+		float GetAmplitude(float waveLength) const override;
 
 	private:
-		double amplitude_;
+		float amplitude_;
 	};
 
 }
