+++ SPTracer/src/Model/Color/SpectralColor.h
@@ -12,18 +12,18 @@ namespace SPTracer
 	public:
 		struct Amplitude
 		{
-			double waveLength;	// length
-			double amplitude;	// amplitude
+			float waveLength;	// length
+			float amplitude;	// amplitude
 		};
 
 		SpectralColor();
 
-		void AddAmplitude(double waveLength, double amplitude);
+		void AddAmplitude(float waveLength, float amplitude);
 		void Init();
-		double GetAmplitude(double waveLength) const override;
+		float GetAmplitude(float waveLength) const override;
 
 	private:
-		static const double WaveLengthAccuracy;
+		static const float WaveLengthAccuracy;
 		std::vector<Amplitude> amplitudes_;
 		bool initialized_ = false;
 	};
